Move files to proper homes
