+++ src/components/Movies/Movies.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import Card from '../Card/Card';
-import movieData from '../movieData';
+import movieData from '../../movieData';
 import './Movies.css';
 
 const Movies = ({movies}) => {
